REFACTOR: GREEN 1st level refactor
